LPS-83427 Skip 'build-ext.gradle' check for Liferay Spring Boot apps

+++ portal-kernel/test/unit/com/liferay/portal/modules/ModulesStructureTest.java
@@ -196,12 +196,14 @@ public class ModulesStructureTest {
 											"plugin\""));
 						}
 
-						Path buildExtGradlePath = dirPath.resolve(
-							"build-ext.gradle");
+						if (!liferaySpringBootDefaultsPlugin) {
+							Path buildExtGradlePath = dirPath.resolve(
+								"build-ext.gradle");
 
-						Assert.assertFalse(
-							"Forbidden " + buildExtGradlePath,
-							Files.deleteIfExists(buildExtGradlePath));
+							Assert.assertFalse(
+								"Forbidden " + buildExtGradlePath,
+								Files.deleteIfExists(buildExtGradlePath));
+						}
 					}
 
 					if (Files.exists(dirPath.resolve("package.json"))) {
